profile photo update

+++ flutter/v1/lib/screens/profile/profile.screen.dart
@@ -60,7 +60,9 @@ class _ProfileScreenState extends State<ProfileScreen> {
 
                         try {
                           /// delete previous file to prevent having unused files in storage.
-                          await ff.deleteFile(ff.user.photoURL);
+                          if (!ff.user.photoURL.isNullOrBlank) {
+                            await ff.deleteFile(ff.user.photoURL);
+                          }
 
                           /// upload picked file,
                           final url = await ff.uploadFile(
